Fixes to remove warnings and error on WIN64 compilations
git-svn-id: https://www.h3d.org:8090/H3DAPI/trunk/H3DAPI@1228 610dfd11-2700-0410-9b68-9ae28eac5983

+++ include/H3D/MField.h
@@ -151,7 +151,7 @@ namespace H3D {
     /// Returns the size of the vector.
     inline unsigned int size() { 
       this->upToDate();
-      return value.size(); 
+      return (unsigned int)value.size(); 
     }
 
     /// Returns the largest possible size of the vector.
@@ -292,7 +292,7 @@ namespace H3D {
                                           unsigned int len, 
                                           int id = 0 ) {
       unsigned int sz = sizeof( value_type );
-      nr_elements = this->value.size();
+      nr_elements = (unsigned int) this->value.size();
       if( len < sz * nr_elements ) {
         return -1;
       }
