+++ src/CoordinateInterpolator2D.cpp
@@ -76,7 +76,7 @@ void CoordinateInterpolator2D::MFValue::update() {
   vector< Vec2f > key_values = static_cast<MFVec2f*>(routes_in[2])->getValue();
   int value_size = 0;
   if( key_size != 0 )
-   value_size = key_values.size() / key_size;
+   value_size = (int) key_values.size() / key_size;
   value.resize( value_size );
 
   if ( key_index >= 0 &&
